refactor: move each command's definition into its own file
cli.ts previously held the npm-shaped arg schema, the default/add command
logic (runPlain/createRun), and every gunshi command object inline — the odd
one out, since sync/init/jsr already kept their behavior in their own files.

Extract the shared flag schema + gunshi plumbing (Ctx, selectorsOf) into
args.ts, move the default/add command (which share createRun) into add.ts,
and have sync.ts/init.ts/jsr-cmd.ts each export their own full command object
(name, description, args, run) next to the logic it calls. cli.ts is now
just imports + gunshi wiring.
